+++ src/common/room.hpp
@@ -6,8 +6,8 @@
 using json = nlohmann::json;
 
 struct Room {
-  size_t room_id = 0;
-  size_t players = 0;
+  uint32_t room_id = 0;
+  uint32_t players = 0;
   GameStatus status = GameStatus::LOBBY;
 };
